+++ src/components/datagrid/fixtures/index.js
@@ -13,3 +13,4 @@ export withConditionalFieldsAndValidations from './comp-with-conditional-compone
 export withLogic from './comp-with-logic';
 export withCollapsibleRowGroups from './comp-with-collapsible-groups';
 export withAllowCalculateOverride from './comp-with-allow-calculate-override';
+export twoWithAllowCalculatedOverride from './two-comp-with-allow-calculate-override';