+++ src/components/html/HTML.js
@@ -92,8 +92,17 @@ export default class HTMLComponent extends Component {
     return super.render(this.renderContent());
   }
 
+  get dataReady() {
+    return this.root?.submissionReady || Promise.resolve();
+  }
+
   attach(element) {
     this.loadRefs(element, { html: 'single' });
+    this.dataReady.then(() => {
+      if (this.refs.html) {
+        this.setContent(this.refs.html, this.content);
+      }
+    });
     return super.attach(element);
   }
 }